TSP-125: user program (#22)
Combine tables User and UserProfile
User program implemented
---------

+++ src/main/java/com/github/sportbot/model/UserProgramId.java
@@ -16,5 +16,5 @@ public class UserProgramId implements Serializable {
     @Column(name = "user_id")
     private Integer userId;
     @Column(name = "exercise_type_id")
-    private Integer exerciseTypeId;
+    private Long exerciseTypeId;
 }
